extract controlled code view from code screen for reuse
