+++ packages/app-design-to-code/code-screen.tsx
@@ -1,136 +1,48 @@
 import React, { useEffect, useState } from "react";
-import { CodeBox, SourceInput } from "@ui/codebox";
 import { Preview } from "@ui/previewer";
-import {
-  EK_GENERATED_CODE_PLAIN,
-  EK_IMAGE_ASSET_REPOSITORY_MAP,
-} from "@core/constant/ek.constant";
-import { repo_assets } from "@design-sdk/core";
 import { assistant as analytics } from "@analytics.bridged.xyz/internal";
-import {
-  FrameworkOption,
-  all_preset_options_map__prod,
-} from "./framework-option";
+import { DesigntoCodeUserOptions } from "./user-options";
 import styled from "@emotion/styled";
-import { make_empty_selection_state_text_content } from "./constants";
-import { format } from "./formatter";
+// import { make_empty_selection_state_text_content } from "./constants";
+
 import copy from "copy-to-clipboard";
 import { PluginSdk } from "@plugin-sdk/app";
 import { CodeScreenFooter } from "./footer-action/code-screen-footer";
-import { CodeOptionsControl } from "./code-options-control";
-import { fromApp, CodeGenRequest } from "./__plugin/events";
-import { useSingleSelection } from "plugin-app";
 
-type DesigntoCodeUserOptions = FrameworkOption;
+import { useSingleSelection } from "plugin-app";
+import { CodeViewWithControl } from "./code-view-with-control";
 
 export function CodeScreen() {
-  const [app, setApp] = useState<string>();
-  const [useroption, setUseroption] = React.useState<DesigntoCodeUserOptions>(
-    all_preset_options_map__prod.flutter_default
-  );
-  const [source, setSource] = useState<SourceInput>();
   const selection = useSingleSelection();
 
-  /** register event listener for events from code thread. */
-  useEffect(() => {
-    window.addEventListener("message", onMessage);
-    return function cleaup() {
-      window.removeEventListener("message", onMessage);
-    };
-  }, [useroption.language]);
+  const [source, setSource] = useState<string>();
+  const [app, setApp] = useState<string>();
+  const [useroption, setUseroption] = useState<DesigntoCodeUserOptions>();
+  // const _make_placeholder = () => {
+  //   return make_empty_selection_state_text_content("empty");
+  // };
 
-  /** post to code thread about target framework change */
-  useEffect(() => {
-    // 1. clear previous result.
-    setSource(undefined);
-    // 2. request new code gen.
-    fromApp({
-      type: "code-gen-request",
-      option: useroption,
-    });
-  }, [useroption.framework, selection]);
-
-  const _make_placeholder = () => {
-    return make_empty_selection_state_text_content("empty");
-  };
-
-  const _make_source = (): SourceInput => {
-    if (typeof source == "string") {
-      if (source && source.length > 0) {
-        return source;
-      }
-    } else {
-      if (source && source.raw.length > 0) {
-        return source;
-      }
-    }
-    return _make_placeholder();
-  };
-
-  const handleSourceInput = ({
-    app,
-    code,
-  }: {
-    app: string;
-    code: SourceInput;
-  }) => {
-    format(app, useroption.language, (s) => {
-      setApp(s);
-    });
-
-    // for SourceInput, we need type checking
-    if (typeof code == "string") {
-      // source input as string
-      format(code, useroption.language, (s) => {
-        setSource(s);
-      });
-    } else {
-      // source input as { raw: string }
-      format(code.raw, useroption.language, (s) => {
-        setSource({
-          raw: s,
-        });
-      });
-    }
-  };
-
-  const onMessage = (ev: MessageEvent) => {
-    const msg = ev.data.pluginMessage;
-    if (msg) {
-      switch (msg.type) {
-        case EK_GENERATED_CODE_PLAIN:
-          handleSourceInput({
-            app: msg.data.app,
-            code: msg.data.code,
-          });
-          // analytics
-          analytics.event_selection_to_code({
-            framework: useroption.framework,
-          });
-
-          break;
-        case EK_IMAGE_ASSET_REPOSITORY_MAP:
-          const imageRepo = msg.data as repo_assets.TransportableImageRepository;
-          repo_assets.ImageHostingRepository.setRepository(imageRepo);
-          break;
-      }
-    } else {
-      // ignore
-    }
-  };
+  // const _make_source = (): SourceInput => {
+  //   if (typeof source == "string") {
+  //     if (source && source.length > 0) {
+  //       return source;
+  //     }
+  //   } else {
+  //     if (source && source.raw.length > 0) {
+  //       return source;
+  //     }
+  //   }
+  //   return _make_placeholder();
+  // };
 
   const onCopyClicked = (e) => {
-    const _code: SourceInput = _make_source();
-    const raw = typeof _code == "string" ? _code : _code.raw;
-    copy(raw);
+    // const _code: SourceInput = _make_source();
+    // const raw = typeof _code == "string" ? _code : _code.raw;
+    copy(source);
     PluginSdk.notifyCopied();
 
     // ANALYTICS
     analytics.event_click_copy_code();
-  };
-
-  const onOptionChange = (op: DesigntoCodeUserOptions) => {
-    setUseroption(op);
   };
 
   return (
@@ -152,43 +64,23 @@ export function CodeScreen() {
           />
         </svg>
       </CopyCodeButton>
-      <CodeWrapper>
-        <CodeOptionsControl
-          // key={JSON.stringify(useroption)} // FIXME: do not uncomment me
-          // initialPreset="react_default" // FIXME: do not uncomment me
-          onUseroptionChange={onOptionChange}
-        />
-        <CodeBox
-          editor="monaco"
-          language={_src_view_language(useroption.framework)}
-          code={_make_source()}
-        />
-      </CodeWrapper>
-
+      <CodeViewWithControl
+        targetid={selection?.id}
+        onGeneration={(app, src) => {
+          setApp(app);
+          setSource(src);
+        }}
+        onUserOptionsChange={setUseroption}
+      />
       <CodeScreenFooter
-        framework={useroption.framework}
+        key={useroption?.framework}
+        framework={useroption?.framework}
         app={app}
         scene={selection?.node as any}
       />
     </div>
   );
 }
-
-/**
- * get language by framework (default) (for code display) (non critical)
- *
- * -- used by code view (for styling only - used by highlight js)
- */
-const _src_view_language = (framework: string): string => {
-  switch (framework) {
-    case "flutter":
-      return "dart";
-    case "react":
-      return "jsx";
-    default:
-      throw `default language for code display on framework "${framework}" is not supported`;
-  }
-};
 
 const CopyCodeButton = styled.div`
   width: 24px;
@@ -198,13 +90,4 @@ const CopyCodeButton = styled.div`
   margin-top: 24px;
   margin-right: 20px;
   cursor: pointer;
-`;
-
-const _VSCODE_DARK_BG = "#1e1e1e";
-
-const CodeWrapper = styled.div`
-  /* 366px is preview(200) + navigation(52+40) + footer btn wrapper(74) height*/
-  height: calc(100vh - 366px);
-  background: ${_VSCODE_DARK_BG};
-  overflow-y: hidden;
 `;
